requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
cssselect2==0.8.0
Django==5.2.3
django-bootstrap-modal-forms==3.0.5
django-braces==1.17.0
django-utils-six==2.0
django-widget-tweaks==1.5.0
html5lib==1.1
idna==3.10
lxml==5.4.0
oscrypto==1.3.0
pillow==11.2.1
pycparser==2.22
pyHanko==0.29.1
pyhanko-certvalidator==0.27.0
pypdf==5.6.0
python-bidi==0.6.6
python-decouple==3.8
PyYAML==6.0.2
reportlab==4.4.2
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
webencodings==0.5.1
xhtml2pdf==0.2.17
